changed some of the pid code scattered throughout the code so that it is putting in a double instead of an int

+++ src/org/TexasTorque/TexasTorque2013/subsystem/drivebase/Drivebase.java
@@ -100,8 +100,8 @@ public class Drivebase
     private synchronized void calcGyroPID()
     {
         desiredGyroAngle = (sensorInput.getGyroAngle() + SmartDashboard.getNumber("azimuth", 0.0));
-        gyroPID.setDesiredValue((int)desiredGyroAngle);
-        double motorOutput = gyroPID.calcPID((int)sensorInput.getGyroAngle());
+        gyroPID.setDesiredValue(desiredGyroAngle);
+        double motorOutput = gyroPID.calcPID(sensorInput.getGyroAngle());
         leftDriveSpeed = motorOutput;
         rightDriveSpeed = -motorOutput;
     }
